+++ .changeset/README.md
@@ -1,37 +1,8 @@
-
 # Changesets
 
-Hello and welcome! This folder has been automatically generated by `@atlaskit/build-releases`, a build tool that works with `bolt` to help you release components from a mono-repository. You can find the full documentation for it [here](https://www.npmjs.com/package/@atlaskit/build-releases)
+Hello and welcome! This folder has been automatically generated by `@changesets/cli`, a build tool that works
+with multi-package repos, or single-package repos to help you version and publish your code. You can
+find the full documentation for it [in our repository](https://github.com/changesets/changesets)
 
-To help you get started though, here are some things you should know about this folder:
-
-## Changesets are automatically generated
-
-Changesets are generated by the `build-releases changeset` command, though it may have been given a new name within your repository. As long as you are following a changeset release flow, you shouldn't have any problems.
-
-## Each changeset is its own folder
-
-We use hashes by default for these folder names to avoid collisions when generating them, but there's no harm that will come from renaming them.
-
-## Changesets are automatically removed
-
-When `build-releases version` or equivalent command is run, all the changeset folders are removed. This is so we only ever use a changeset once. This makes this a very bad place to store any other information.
-
-## Changesets come in two parts
-
-You should treat these parts quite differently:
-
-- `changes.md` is a file you should feel free to edit as much as you want. It will be prepended to your changelog when you next run your version command.
-- `changes.json` is a file that includes information about releases, what should be versioned by the version command. We strongly recommend against editing this directly, as you may make a new changeset that puts your bolt repository into an invalid state.
-
-## I want to edit the information in a `changes.json` - how do I do it safely?
-
-The best option is to make a new changeset using the changeset command, copy over the `changes.md`, then delete the old changeset.
-
-## Can I rename the folder for my changeset?
-
-Absolutely! We need unique hashes to make changesets play nicely with git, but changing your folder from our hash to your own name isn't going to cause any problems.
-
-## Can I manually delete changesets?
-
-You can, but you should be aware this will remove the intent to release communicated by the changeset, and should be done with caution.
+We have a quick list of common questions to get you started engaging with this project in
+[our documentation](https://github.com/changesets/changesets/blob/master/docs/common-questions.md)
