+++ .changeset/config.json
@@ -1,0 +1,7 @@
+{
+  "$schema": "https://unpkg.com/@changesets/config@0.2.1/schema.json",
+  "changelog": "./getChangelogEntry",
+  "commit": false,
+  "linked": [],
+  "access": "public"
+}
