+++ packages/build-field-types/package.json
@@ -29,12 +29,9 @@
     "builtin-modules": "^3.0.0",
     "chalk": "^2.4.2",
     "dataloader": "^1.4.0",
-    "diff": "^4.0.1",
-    "fast-deep-equal": "^2.0.1",
     "fs-extra": "^7.0.0",
     "globby": "^9.1.0",
     "inquirer": "^6.2.0",
-    "install-packages": "^0.2.5",
     "jest-worker": "^24.6.0",
     "lodash.isempty": "^4.4.0",
     "magic-string": "^0.25.1",
@@ -60,7 +57,6 @@
     "@babel/preset-flow": "^7.0.0",
     "@babel/runtime": "^7.4.3",
     "fixturez": "^1.1.0",
-    "jest-in-case": "^1.0.2",
     "react": "^16.8.6",
     "rimraf": "^2.6.2",
     "spawndamnit": "^2.0.0"
