+++ test-projects/access-control/package.json
@@ -35,6 +35,9 @@
     "mocha-junit-reporter": "^1.21.0",
     "mocha-multi-reporters": "^1.1.7",
     "pino-colada": "^1.4.4",
-    "start-server-and-test": "^1.10.2"
+    "start-server-and-test": "^1.10.2",
+    "inflection": "^1.12.0",
+    "execa": "^2.0.4"
+
   }
 }
