+++ test-projects/login/package.json
@@ -38,6 +38,7 @@
     "mocha-junit-reporter": "^1.21.0",
     "mocha-multi-reporters": "^1.1.7",
     "pino-colada": "^1.4.4",
-    "start-server-and-test": "^1.10.2"
+    "start-server-and-test": "^1.10.2",
+    "execa": "^2.0.4"
   }
 }
