+++ website/gatsby-config.js
@@ -79,9 +79,6 @@ async function getGatsbyConfig() {
         resolve: `gatsby-mdx`,
         options: {
           extensions: ['.mdx', '.md'],
-          globalScope: `import { Props } from '${require.resolve('./src/components/props')}'
-          export default { Props }
-          `,
           defaultLayouts: {
             default: require.resolve('./src/components/mdx-renderer.js'),
           },
