+++ website/gatsby-node.js
@@ -108,10 +108,6 @@ exports.onCreateBabelConfig = ({ actions, stage }) => {
     name: `@babel/preset-flow`,
     stage,
   });
-  actions.setBabelPlugin({
-    name: `babel-plugin-extract-react-types`,
-    stage,
-  });
 };
 
 const getEditUrl = absPath =>
