--- website/src/components/props.js
@@ -1,6 +1,0 @@
-import * as React from 'react';
-import PrettyProps from 'pretty-proptypes';
-
-export let Props = ({ component, ...props }) => {
-  return <PrettyProps props={{ type: 'program', component: component.___types }} {...props} />;
-};
